fixed bugs related to delays in the UI when connecting iPad and iPhone. Fixed FlexSensor's sprite & pins. Fixed some warnings related to 64 bit architecture

+++ TangoHappsClient/classes/THClientDownloadViewController.h
@@ -57,7 +57,7 @@ const float kIconInstallationUpdateFrequency;
     
     NSTimer * installationProgressTimer;
     NSTimeInterval timeWhenInstallationStarted;
-    float installationUpdateRate;
+    CGFloat installationUpdateRate;
 }
 
 @property (weak, nonatomic) IBOutlet UILabel *instructionsLabel;
